map/pin api2

+++ lib/map/model/request_shared_map/request_shared_map_model.dart
@@ -1,10 +1,6 @@
-import 'dart:ui';
-
-import 'package:caps_2/models/map_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'request_shared_map_model.freezed.dart';
-
 part 'request_shared_map_model.g.dart';
 
 @freezed
@@ -12,8 +8,8 @@ class RequestSharedMapModel with _$RequestSharedMapModel {
   const factory RequestSharedMapModel({
     required String title,
     required String color,
-    required int lat,
-    required int lon,
+    required double lat,
+    required double lon,
     required String selectedDate,
   }) = _RequestSharedMapModel;
